Connect backend to database successfully
Firstly initialize the database in HeidiSQL.
Execute init.sql to create the database and table.

Then run backend with the following command:
```
node app.js
```
and visit http://localhost:3000 in browser to check
if admin table is displayed.

Now backend can connect to database successfully.
The next step is to implement the API for the
frontend to access the database.

Signed-off-by: Liu Jing <liujing522@163.com>